✨ refactor(curation): 쉬머(Shimmer) 및 스켈레톤 UI 컴포넌트 구조 고도화
- **공통 레이아웃 유틸리티 정리**
    - `CurationConstants`, `Responsive` 등 미사용 및 레거시 레이아웃 상수/헬퍼 파일 삭제
    - 여러 곳에 흩어져 있던 `shimmer` 관련 로직을 `ui/effect/skeleton` 패키지로 통합

- **쉬머(Shimmer) 애니메이션 시스템 개선**
    - `Modifier.shimmer()` 확장 함수 고도화: `Shape` 지원 및 `drawWithCache`를 통한 성능 최적화
    - 테마와 연동된 색상 체계 도입: `grayShimmerColors`(범용), `pinkShimmerColors`(큐레이션 상세용) 정의
    - `ShimmerBrush.kt`를 `Shimmer.kt`로 파일명을 변경하고 구현 방식을 Brush 주입에서 Modifier 기반으로 전환

- **스켈레톤(Skeleton) 컴포넌트 리팩터링**
    - `SkeletonBox` 컴포넌트가 `Modifier.shimmer`를 내부적으로 사용하도록 간소화
    - 기존 `shimmerBrush` 파라미터를 `colors: List<Color>` 형태로 변경하여 유연성 확보
    - 사용하지 않는 레거시 스켈레톤 코드(`ShimmerSkeleton.kt`) 삭제 및 KDoc 주석을 통한 사용 가이드 추가
