+++ feature/curation/src/main/java/com/linku/curation/ui/effect/skeleton/SkeletonBox.kt
@@ -1,58 +1,60 @@
 package com.linku.curation.ui.effect.skeleton
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
+import androidx.compose.ui.graphics.Shape
 
+/**
+ * 쉬머 애니메이션이 적용된 스켈레톤 로딩 플레이스홀더 Box.
+ *
+ * @param shape 스켈레톤 영역의 모양 (원형, 라운드, 사각형 등 자유롭게 지정)
+ * @param colors 쉬머 그라데이션 색상 목록 (기본값: [grayShimmerColors])
+ */
 @Composable
 fun SkeletonBox(
     modifier: Modifier = Modifier,
     shape: Shape = RectangleShape,
-    shimmerBrush: Brush
+    colors: List<Color> = grayShimmerColors()
 ) {
-    Box(
-        modifier = modifier
-            .clip(shape)
-            .background(shimmerBrush)
-    )
+    Box(modifier = modifier.shimmer(shape = shape, colors = colors))
 }
 
 /**
  * 사용법
- * 카드 스켈레톤(라운드가 있다면?) + 일반 스켈레톤
+ *
+ * 카드 스켈레톤(라운드)
  * SkeletonBox(
- *     modifier = Modifier
- *         .size(346.scaler, 432.scaler),
- *     shape = RoundedCornerShape(24.scaler),
- *     shimmerBrush = grayShimmerBrush()
+ *     modifier = Modifier.size(346.scaler, 432.scaler),
+ *     shape = RoundedCornerShape(24.scaler)
  * )
  *
  *
- * 핑크 쉬머(큐레이션 디테일 화면에 사용?)
+ * 원형 스켈레톤(프로필 등)
+ * SkeletonBox(
+ *     modifier = Modifier.size(48.scaler),
+ *     shape = CircleShape
+ * )
+ *
+ *
+ * 핑크 쉬머(큐레이션 디테일 화면)
  * SkeletonBox(
  *     modifier = Modifier
  *         .fillMaxWidth()
  *         .height(180.scaler),
  *     shape = RoundedCornerShape(20.scaler),
- *     shimmerBrush = pinkShimmerBrush()
+ *     colors = pinkShimmerColors()
  * )
  *
  *
- * //텍스트 라인
+ * 텍스트 라인
  * SkeletonBox(
  *     modifier = Modifier
  *         .height(14.scaler)
  *         .fillMaxWidth(0.6f),
- *     shape = RoundedCornerShape(4.scaler),
- *     shimmerBrush = grayShimmerBrush()
+ *     shape = RoundedCornerShape(4.scaler)
  * )
- *
- *
- *
  *
  * */
