--- feature/curation/src/main/java/com/linku/curation/ui/util/CurationConstants.kt
@@ -1,24 +1,0 @@
-package com.linku.curation.ui.util
-
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-
-/**
- * 큐레이션 화면 공통 레이아웃 상수 (dp 값)
- * 사용 시 .scaler 적용 필요
- * 큐레이션에서만 사용함.
- */
-object CurationConstants {
-    // 카드 페이저 위치 (화면 상단에서부터)
-    const val CARD_TOP_OFFSET_VALUE = 174
-
-    // 카드 좌우 패딩
-    const val CARD_HORIZONTAL_PADDING_VALUE = 33
-
-    // 카드 크기
-    const val CARD_WIDTH_VALUE = 346
-    const val CARD_HEIGHT_VALUE = 432
-
-    // 디테일 카드 높이
-    const val DETAIL_CARD_HEIGHT_VALUE = 350
-}
