--- feature/curation/src/main/java/com/linku/curation/ui/util/ShimmerSkeleton.kt
@@ -1,24 +1,0 @@
-//package com.linku.curation.ui.util
-//
-//import androidx.compose.foundation.background
-//import androidx.compose.foundation.layout.Box
-//import androidx.compose.runtime.Composable
-//import androidx.compose.ui.Modifier
-//import androidx.compose.ui.graphics.Brush
-//
-//
-//@Composable
-//fun ShimmerSkeleton(modifier: Modifier = Modifier) {
-//    Box(
-//        modifier = modifier
-//            .shimmer()
-//            .background(
-//                Brush.verticalGradient(
-//                    colors = listOf(
-//                        SkeletonStart,
-//                        SkeletonEnd
-//                    )
-//                )
-//            )
-//    )
-//}
